csr_kernels
