+++ cuda/test/CMakeLists.txt
@@ -2,7 +2,6 @@ include(${PROJECT_SOURCE_DIR}/cmake/create_test.cmake)
 
 add_subdirectory(base)
 add_subdirectory(components)
-add_subdirectory(matrix)
 add_subdirectory(reorder)
 add_subdirectory(solver)
 add_subdirectory(utils)
