--- cuda/test/matrix/CMakeLists.txt
@@ -1,1 +1,0 @@
-ginkgo_create_test(csr_kernels)
