+++ hip/test/matrix/CMakeLists.txt
@@ -1,4 +1,3 @@
-ginkgo_create_hip_test(csr_kernels)
 ginkgo_create_hip_test(fbcsr_kernels)
 if (hipfft_FOUND)
     ginkgo_create_hip_test(fft_kernels)
