+++ omp/test/matrix/CMakeLists.txt
@@ -1,2 +1,1 @@
-ginkgo_create_test(csr_kernels)
 ginkgo_create_test(fbcsr_kernels)
